done with crud apis

+++ src/main/java/com/Learning/Project/service/EmployeeService.java
@@ -27,8 +27,40 @@ public class EmployeeService {
         return employeeRepository.save(employee);
     }
 
-    public void deleteEmployee(String id) {
-        employeeRepository.deleteById(id);
+
+    public boolean deleteEmployee(String id) {
+        Optional<Employee> employee = employeeRepository.findById(id);
+        if (employee.isPresent()) {
+            employeeRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+    public Optional<Employee> updateEmployee(String id, Employee updatedEmployee) {
+        return employeeRepository.findById(id).map(existingEmployee -> {
+            if (updatedEmployee.getName() != null) {
+                existingEmployee.setName(updatedEmployee.getName());
+            }
+            if (updatedEmployee.getPosition() != null) {
+                existingEmployee.setPosition(updatedEmployee.getPosition());
+            }
+            if (updatedEmployee.getDepartment() != null) {
+                existingEmployee.setDepartment(updatedEmployee.getDepartment());
+            }
+            if (updatedEmployee.getEmail() != null) {
+                existingEmployee.setEmail(updatedEmployee.getEmail());
+            }
+            if (updatedEmployee.getCountryCode() != null) {
+                existingEmployee.setCountry_code(updatedEmployee.getCountryCode());
+            }
+            if (updatedEmployee.getPhoneNumber() != null) {
+                existingEmployee.setPhone_number(updatedEmployee.getPhoneNumber());
+            }
+            existingEmployee.setUpdatedAt(updatedEmployee.getUpdatedAt());
+
+            return employeeRepository.save(existingEmployee);
+        });
     }
 
     public EmployeeResponseDTO convertToDTO(Employee employee) {
@@ -40,7 +72,6 @@ public class EmployeeService {
         dto.setCountry_code(employee.getCountryCode());
         dto.setCountry_code(employee.getCountryCode());
         dto.setEmail(employee.getEmail());
-
 
 
         return dto;
